# specs/001-food-calorie-mvp/contracts/api.yaml
openapi: 3.1.0
info:
  title: CalSnap MVP API
  version: 1.0.0
  description: >
    Local API contract for the CalSnap food calorie MVP. All protected endpoints require
    a bearer token issued by the API.
servers:
  - url: http://localhost:3000
tags:
  - name: Auth
  - name: Analysis
  - name: History
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        retryable:
          type: boolean
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
    AuthSessionResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    AnalysisResult:
      type: object
      required:
        - foodName
        - portionDescription
        - caloriesKcal
        - proteinGrams
        - carbsGrams
        - fatGrams
        - confidenceLabel
        - confidenceReason
        - estimateDisclaimer
        - retryRecommended
      properties:
        foodName:
          type: string
        portionDescription:
          type: string
        caloriesKcal:
          type: number
          minimum: 0
        proteinGrams:
          type: number
          minimum: 0
        carbsGrams:
          type: number
          minimum: 0
        fatGrams:
          type: number
          minimum: 0
        confidenceLabel:
          type: string
          enum: [high, medium, low]
        confidenceReason:
          type: string
        estimateDisclaimer:
          type: string
        retryRecommended:
          type: boolean
        retryReason:
          type: string
          nullable: true
    AnalysisSuccessResponse:
      type: object
      required: [status, recordId, result]
      properties:
        status:
          type: string
          const: succeeded
        recordId:
          type: string
        result:
          $ref: '#/components/schemas/AnalysisResult'
    AnalysisRetryResponse:
      type: object
      required: [status, result]
      properties:
        status:
          type: string
          const: needs_retry
        result:
          type: object
          required:
            - confidenceLabel
            - confidenceReason
            - estimateDisclaimer
            - retryRecommended
            - retryReason
          properties:
            confidenceLabel:
              type: string
              enum: [low]
            confidenceReason:
              type: string
            estimateDisclaimer:
              type: string
            retryRecommended:
              type: boolean
              const: true
            retryReason:
              type: string
    HistoryListItem:
      type: object
      required:
        - id
        - foodName
        - portionDescription
        - caloriesKcal
        - confidenceLabel
        - createdAt
      properties:
        id:
          type: string
        foodName:
          type: string
        portionDescription:
          type: string
        caloriesKcal:
          type: number
        confidenceLabel:
          type: string
          enum: [high, medium, low]
        createdAt:
          type: string
          format: date-time
    HistoryDetail:
      type: object
      required:
        - id
        - foodName
        - portionDescription
        - caloriesKcal
        - proteinGrams
        - carbsGrams
        - fatGrams
        - confidenceLabel
        - confidenceReason
        - estimateDisclaimer
        - createdAt
      properties:
        id:
          type: string
        foodName:
          type: string
        portionDescription:
          type: string
        caloriesKcal:
          type: number
        proteinGrams:
          type: number
        carbsGrams:
          type: number
        fatGrams:
          type: number
        confidenceLabel:
          type: string
          enum: [high, medium, low]
        confidenceReason:
          type: string
        estimateDisclaimer:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created and signed in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: Fetch the current signed-in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Session revoked
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analysis:
    post:
      tags: [Analysis]
      summary: Analyze one uploaded image containing a single food or beverage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Structured result or retry guidance
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AnalysisSuccessResponse'
                  - $ref: '#/components/schemas/AnalysisRetryResponse'
        '400':
          description: Invalid upload or unsupported image payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Analysis service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /history:
    get:
      tags: [History]
      summary: List saved records for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's saved records
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryListItem'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /history/{recordId}:
    get:
      tags: [History]
      summary: Fetch one saved record owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saved record detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDetail'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Record not found or not owned by current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [History]
      summary: Delete one saved record owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Record deleted
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Record not found or not owned by current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'